requirements.txt
# requirements.txt
# Core
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
python-dotenv==1.0.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.28
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Telegram
python-telegram-bot==20.8

# Task Queue
celery==5.3.6
redis==5.0.2

# HTTP Client
aiohttp==3.9.3
requests==2.31.0

# AI Integration
openai==1.28.0
anthropic==0.20.0

# Payment Processing
razorpay==1.4.1

# Parsing
python-docx==1.1.0
PyPDF2==3.0.1

# Utils
tenacity==8.2.3
ujson==5.10.0
pyjwt==2.8.0
cryptography==42.0.5

# Logging
structlog==24.1.0

# Add this to your requirements.txt
razorpay==1.4.1

sqlalchemy_utils